Checkpoint

+++ src/databricks_sql/transform/05_create_fact_orders.sql
@@ -3,9 +3,9 @@ DROP TABLE IF EXISTS fact_orders;
 
 CREATE TABLE fact_orders (
     order_id STRING,
-    customer_id INT,
-    product_id INT,
-    date_id INT,
+    customer_id BIGINT,
+    product_id BIGINT,
+    date_id BIGINT,
     price DECIMAL(10,2),
     passenger_count INT,
     CONSTRAINT fk_customer FOREIGN KEY (customer_id) REFERENCES dim_customers(customer_id),
